+++ json_serializable/lib/src/decode_helper.dart
@@ -32,9 +32,11 @@ mixin DecodeHelper implements HelperCore {
     final buffer = StringBuffer();
 
     final mapType = config.anyMap ? 'Map' : 'Map<String, dynamic>';
-    buffer.write('$targetClassReference '
-        '${prefix}FromJson${genericClassArgumentsImpl(withConstraints: true)}'
-        '($mapType json');
+    buffer.write(
+      '$targetClassReference '
+      '${prefix}FromJson${genericClassArgumentsImpl(withConstraints: true)}'
+      '($mapType json',
+    );
 
     if (config.genericArgumentFactories) {
       for (var arg in element.typeParameters) {
@@ -53,30 +55,41 @@ mixin DecodeHelper implements HelperCore {
 
     final fromJsonLines = <String>[];
 
-    String deserializeFun(String paramOrFieldName,
-            {ParameterElement? ctorParam}) =>
-        _deserializeForField(accessibleFields[paramOrFieldName]!,
-            ctorParam: ctorParam);
+    String deserializeFun(
+      String paramOrFieldName, {
+      FormalParameterElement? ctorParam,
+    }) => _deserializeForField(
+      accessibleFields[paramOrFieldName]!,
+      ctorParam: ctorParam,
+    );
 
     final data = _writeConstructorInvocation(
       element,
       config.constructor,
       accessibleFields.keys,
       accessibleFields.values
-          .where((fe) => element.lookUpSetter(fe.name, element.library) != null)
-          .map((fe) => fe.name)
+          .where(
+            (fe) =>
+                element.lookUpSetter(
+                  name: fe.name!,
+                  library: element.library,
+                ) !=
+                null,
+          )
+          .map((fe) => fe.name!)
           .toList(),
       unavailableReasons,
       deserializeFun,
     );
 
     final checks = _checkKeys(
-      accessibleFields.values
-          .where((fe) => data.usedCtorParamsAndFields.contains(fe.name)),
+      accessibleFields.values.where(
+        (fe) => data.usedCtorParamsAndFields.contains(fe.name),
+      ),
     ).toList();
 
     if (config.checked) {
-      final classLiteral = escapeDartString(element.name);
+      final classLiteral = escapeDartString(element.name!);
 
       final sectionBuffer = StringBuffer()
         ..write('''
@@ -96,9 +109,7 @@ mixin DecodeHelper implements HelperCore {
           ..write('''
     \$checkedConvert($safeName, (v) => ''')
           ..write('val.$fieldName = ')
-          ..write(
-            _deserializeForField(fieldValue, checkedProperty: true),
-          );
+          ..write(_deserializeForField(fieldValue, checkedProperty: true));
 
         final readValueFunc = jsonKeyFor(fieldValue).readValueFunctionName;
         if (readValueFunc != null) {
@@ -111,9 +122,11 @@ mixin DecodeHelper implements HelperCore {
       sectionBuffer.write('''\n    return val;
   }''');
 
-      final fieldKeyMap = Map.fromEntries(data.usedCtorParamsAndFields
-          .map((k) => MapEntry(k, nameAccess(accessibleFields[k]!)))
-          .where((me) => me.key != me.value));
+      final fieldKeyMap = Map.fromEntries(
+        data.usedCtorParamsAndFields
+            .map((k) => MapEntry(k, nameAccess(accessibleFields[k]!)))
+            .where((me) => me.key != me.value),
+      );
 
       String fieldKeyMapArg;
       if (fieldKeyMap.isEmpty) {
@@ -130,7 +143,8 @@ mixin DecodeHelper implements HelperCore {
     } else {
       fromJsonLines.addAll(checks);
 
-      final sectionBuffer = StringBuffer()..write('''
+      final sectionBuffer = StringBuffer()
+        ..write('''
   ${data.content}''');
       for (final field in data.fieldsToSet) {
         sectionBuffer
@@ -162,7 +176,7 @@ mixin DecodeHelper implements HelperCore {
     final args = <String>[];
 
     String constantList(Iterable<FieldElement> things) =>
-        'const ${jsonLiteralAsDart(things.map(nameAccess).toList())}';
+        'const ${jsonLiteralAsDart(things.map<String>(nameAccess).toList())}';
 
     if (config.disallowUnrecognizedKeys) {
       final allowKeysLiteral = constantList(accessibleFields);
@@ -170,8 +184,9 @@ mixin DecodeHelper implements HelperCore {
       args.add('allowedKeys: $allowKeysLiteral');
     }
 
-    final requiredKeys =
-        accessibleFields.where((fe) => jsonKeyFor(fe).required).toList();
+    final requiredKeys = accessibleFields
+        .where((fe) => jsonKeyFor(fe).required)
+        .toList();
     if (requiredKeys.isNotEmpty) {
       final requiredKeyLiteral = constantList(requiredKeys);
 
@@ -246,7 +261,7 @@ bool _isBuiltInType(DartType type) => type.isDartAsyncFuture ||
   /// setter.
   String _deserializeForField(
     FieldElement field, {
-    ParameterElement? ctorParam,
+    FormalParameterElement? ctorParam,
     bool checkedProperty = false,
   }) {
     final jsonKeyName = safeNameAccess(field);
@@ -257,11 +272,7 @@ bool _isBuiltInType(DartType type) => type.isDartAsyncFuture ||
     final readValueFunc = jsonKey.readValueFunctionName;
 
     String deserialize(String expression) => contextHelper
-        .deserialize(
-          targetType,
-          expression,
-          defaultValue: defaultValue,
-        )
+        .deserialize(targetType, expression, defaultValue: defaultValue)
         .toString();
 
     String value;
@@ -269,8 +280,9 @@ bool _isBuiltInType(DartType type) => type.isDartAsyncFuture ||
       if (config.checked) {
         value = deserialize('v');
         if (!checkedProperty) {
-          final readValueBit =
-              readValueFunc == null ? '' : ',readValue: $readValueFunc,';
+          final readValueBit = readValueFunc == null
+              ? ''
+              : ',readValue: $readValueFunc,';
           value = '\$checkedConvert($jsonKeyName, (v) => $value$readValueBit)';
         }
       } else {
@@ -292,9 +304,11 @@ bool _isBuiltInType(DartType type) => type.isDartAsyncFuture ||
 
     if (defaultValue != null) {
       if (jsonKey.disallowNullValue && jsonKey.required) {
-        log.warning('The `defaultValue` on field `${field.name}` will have no '
-            'effect because both `disallowNullValue` and `required` are set to '
-            '`true`.');
+        log.warning(
+          'The `defaultValue` on field `${field.name}` will have no '
+          'effect because both `disallowNullValue` and `required` are set to '
+          '`true`.',
+        );
       }
     }
     return value;
@@ -317,21 +331,22 @@ _ConstructorData _writeConstructorInvocation(
   Iterable<String> availableConstructorParameters,
   Iterable<String> writableFields,
   Map<String, String> unavailableReasons,
-  String Function(String paramOrFieldName, {ParameterElement ctorParam})
-      deserializeForField,
+  String Function(String paramOrFieldName, {FormalParameterElement ctorParam})
+  deserializeForField,
 ) {
   final className = classElement.name;
 
   final ctor = constructorByName(classElement, constructorName);
 
   final usedCtorParamsAndFields = <String>{};
-  final constructorArguments = <ParameterElement>[];
-  final namedConstructorArguments = <ParameterElement>[];
+  final constructorArguments = <FormalParameterElement>[];
+  final namedConstructorArguments = <FormalParameterElement>[];
 
-  for (final arg in ctor.parameters) {
+  for (final arg in ctor.formalParameters) {
     if (!availableConstructorParameters.contains(arg.name)) {
       if (arg.isRequired) {
-        var msg = 'Cannot populate the required constructor '
+        var msg =
+            'Cannot populate the required constructor '
             'argument: ${arg.name}.';
 
         final additionalInfo = unavailableReasons[arg.name];
@@ -352,12 +367,13 @@ _ConstructorData _writeConstructorInvocation(
     } else {
       constructorArguments.add(arg);
     }
-    usedCtorParamsAndFields.add(arg.name);
+    usedCtorParamsAndFields.add(arg.name!);
   }
 
   // fields that aren't already set by the constructor and that aren't final
-  final remainingFieldsForInvocationBody =
-      writableFields.toSet().difference(usedCtorParamsAndFields);
+  final remainingFieldsForInvocationBody = writableFields.toSet().difference(
+    usedCtorParamsAndFields,
+  );
 
   final constructorExtra = constructorName.isEmpty ? '' : '.$constructorName';
 
@@ -367,16 +383,24 @@ _ConstructorData _writeConstructorInvocation(
       '${genericClassArguments(classElement, false)}'
       '$constructorExtra(',
     )
-    ..writeAll(constructorArguments.map((paramElement) {
-      final content =
-          deserializeForField(paramElement.name, ctorParam: paramElement);
-      return '      $content,\n';
-    }))
-    ..writeAll(namedConstructorArguments.map((paramElement) {
-      final value =
-          deserializeForField(paramElement.name, ctorParam: paramElement);
-      return '      ${paramElement.name}: $value,\n';
-    }))
+    ..writeAll(
+      constructorArguments.map((paramElement) {
+        final content = deserializeForField(
+          paramElement.name!,
+          ctorParam: paramElement,
+        );
+        return '      $content,\n';
+      }),
+    )
+    ..writeAll(
+      namedConstructorArguments.map((paramElement) {
+        final value = deserializeForField(
+          paramElement.name!,
+          ctorParam: paramElement,
+        );
+        return '      ${paramElement.name!}: $value,\n';
+      }),
+    )
     ..write(')');
 
   usedCtorParamsAndFields.addAll(remainingFieldsForInvocationBody);