+++ json_serializable/lib/src/field_helpers.dart
@@ -3,12 +3,8 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:analyzer/dart/element/element.dart';
-import 'package:analyzer/src/dart/element/element.dart' // ignore: implementation_imports
-    show
-        InterfaceElementImpl;
 import 'package:analyzer/src/dart/element/inheritance_manager3.dart' // ignore: implementation_imports
-    show
-        InheritanceManager3;
+    show InheritanceManager3;
 import 'package:source_gen/source_gen.dart';
 
 import 'utils.dart';
@@ -18,7 +14,7 @@ class _FieldSet implements Comparable<_FieldSet> {
   final FieldElement sortField;
 
   _FieldSet._(this.field, this.sortField)
-      : assert(field.name == sortField.name);
+    : assert(field.name == sortField.name);
 
   factory _FieldSet(FieldElement? classField, FieldElement? superField) {
     // At least one of these will != null, perhaps both.
@@ -29,8 +25,10 @@ class _FieldSet implements Comparable<_FieldSet> {
 
     // Prefer the field that's annotated with `JsonKey`, if any.
     // If not, use the class field.
-    final fieldHasJsonKey =
-        fields.firstWhere(hasJsonKeyAnnotation, orElse: () => fields.first);
+    final fieldHasJsonKey = fields.firstWhere(
+      hasJsonKeyAnnotation,
+      orElse: () => fields.first,
+    );
 
     return _FieldSet._(fieldHasJsonKey, sortField);
   }
@@ -39,22 +37,19 @@ class _FieldSet implements Comparable<_FieldSet> {
   int compareTo(_FieldSet other) => _sortByLocation(sortField, other.sortField);
 
   static int _sortByLocation(FieldElement a, FieldElement b) {
-    final checkerA = TypeChecker.fromStatic(
-      (a.enclosingElement3 as InterfaceElement).thisType,
-    );
+    final checkerA = TypeChecker.fromStatic(a.enclosingElement.thisType);
 
-    if (!checkerA.isExactly(b.enclosingElement3)) {
+    if (!checkerA.isExactly(b.enclosingElement)) {
       // in this case, you want to prioritize the enclosingElement that is more
       // "super".
 
-      if (checkerA.isAssignableFrom(b.enclosingElement3)) {
+      if (checkerA.isAssignableFrom(b.enclosingElement)) {
         return -1;
       }
 
-      final checkerB = TypeChecker.fromStatic(
-          (b.enclosingElement3 as InterfaceElement).thisType);
+      final checkerB = TypeChecker.fromStatic(b.enclosingElement.thisType);
 
-      if (checkerB.isAssignableFrom(a.enclosingElement3)) {
+      if (checkerB.isAssignableFrom(a.enclosingElement)) {
         return 1;
       }
     }
@@ -63,9 +58,9 @@ class _FieldSet implements Comparable<_FieldSet> {
     /// preference for the getter if it's defined.
     int offsetFor(FieldElement e) {
       if (e.isSynthetic) {
-        return (e.getter ?? e.setter)!.nameOffset;
+        return (e.getter ?? e.setter)!.firstFragment.nameOffset ?? 0;
       }
-      return e.nameOffset;
+      return e.firstFragment.nameOffset ?? 0;
     }
 
     return offsetFor(a).compareTo(offsetFor(b));
@@ -79,38 +74,41 @@ List<FieldElement> createSortedFieldSet(ClassElement element) {
   // Get all of the fields that need to be assigned
   // TODO: support overriding the field set with an annotation option
   final elementInstanceFields = Map.fromEntries(
-      element.fields.where((e) => !e.isStatic).map((e) => MapEntry(e.name, e)));
+    element.fields.where((e) => !e.isStatic).map((e) => MapEntry(e.name, e)),
+  );
 
   final inheritedFields = <String, FieldElement>{};
   final manager = InheritanceManager3();
 
-  for (final v in manager
-      .getInheritedConcreteMap2(element as InterfaceElementImpl)
-      .values) {
+  for (final v in manager.getInheritedConcreteMap(element).values) {
     assert(v is! FieldElement);
-    if (_dartCoreObjectChecker.isExactly(v.enclosingElement3)) {
+    if (_dartCoreObjectChecker.isExactly(v.enclosingElement!)) {
       continue;
     }
 
-    if (v is PropertyAccessorElement &&
-        (v as PropertyAccessorElement).isGetter) {
-      assert((v as PropertyAccessorElement).variable2 is FieldElement);
-      final variable = (v as PropertyAccessorElement).variable2 as FieldElement;
+    if (v is GetterElement) {
+      final variable = v.variable as FieldElement;
       assert(!inheritedFields.containsKey(variable.name));
-      inheritedFields[variable.name] = variable;
+      inheritedFields[variable.name!] = variable;
     }
   }
 
   // Get the list of all fields for `element`
-  final allFields =
-      elementInstanceFields.keys.toSet().union(inheritedFields.keys.toSet());
+  final allFields = elementInstanceFields.keys.toSet().union(
+    inheritedFields.keys.toSet(),
+  );
 
-  final fields = allFields
-      .map((e) => _FieldSet(elementInstanceFields[e], inheritedFields[e]))
-      .toList()
-    ..sort();
+  final fields =
+      allFields
+          .map((e) => _FieldSet(elementInstanceFields[e], inheritedFields[e]))
+          .toList()
+        ..sort();
 
   return fields.map((fs) => fs.field).toList(growable: false);
 }
 
-const _dartCoreObjectChecker = TypeChecker.fromRuntime(Object);
+const _dartCoreObjectChecker = TypeChecker.typeNamed(
+  Object,
+  inPackage: 'core',
+  inSdk: true,
+);