+++ json_serializable/lib/src/generator_helper.dart
@@ -23,12 +23,9 @@ class GeneratorHelper extends HelperCore with EncodeHelper, DecodeHelper {
     ClassElement element,
     ConstantReader annotation,
   ) : super(
-            element,
-            mergeConfig(
-              _generator.config,
-              annotation,
-              classElement: element,
-            ));
+        element,
+        mergeConfig(_generator.config, annotation, classElement: element),
+      );
 
   @override
   void addMember(String memberContent) {
@@ -63,18 +60,19 @@ class GeneratorHelper extends HelperCore with EncodeHelper, DecodeHelper {
       (map, field) {
         final jsonKey = jsonKeyFor(field);
         if (!field.isPublic && !jsonKey.explicitYesFromJson) {
-          unavailableReasons[field.name] = 'It is assigned to a private field.';
+          unavailableReasons[field.name!] =
+              'It is assigned to a private field.';
         } else if (field.getter == null) {
           assert(field.setter != null);
-          unavailableReasons[field.name] =
+          unavailableReasons[field.name!] =
               'Setter-only properties are not supported.';
           log.warning('Setters are ignored: ${element.name}.${field.name}');
         } else if (jsonKey.explicitNoFromJson) {
-          unavailableReasons[field.name] =
+          unavailableReasons[field.name!] =
               'It is assigned to a field not meant to be used in fromJson.';
         } else {
           assert(!map.containsKey(field.name));
-          map[field.name] = field;
+          map[field.name!] = field;
         }
 
         return map;
@@ -93,40 +91,37 @@ class GeneratorHelper extends HelperCore with EncodeHelper, DecodeHelper {
 
       // Need to add candidates BACK even if they are not used in the factory if
       // they are forced to be used for toJSON
-      for (var candidate in sortedFields.where((element) =>
-          jsonKeyFor(element).explicitYesToJson &&
-          !fieldsToUse.contains(element))) {
+      for (var candidate in sortedFields.where(
+        (element) =>
+            jsonKeyFor(element).explicitYesToJson &&
+            !fieldsToUse.contains(element),
+      )) {
         fieldsToUse.add(candidate);
       }
 
       // Need the fields to maintain the original source ordering
       fieldsToUse.sort(
-          (a, b) => sortedFields.indexOf(a).compareTo(sortedFields.indexOf(b)));
+        (a, b) => sortedFields.indexOf(a).compareTo(sortedFields.indexOf(b)),
+      );
 
       accessibleFieldSet = fieldsToUse.toSet();
     }
 
     accessibleFieldSet
-      ..removeWhere(
-        (element) => jsonKeyFor(element).explicitNoToJson,
-      )
-
+      ..removeWhere((element) => jsonKeyFor(element).explicitNoToJson)
       // Check for duplicate JSON keys due to colliding annotations.
       // We do this now, since we have a final field list after any pruning done
       // by `_writeCtor`.
-      ..fold(
-        <String>{},
-        (Set<String> set, fe) {
-          final jsonKey = nameAccess(fe);
-          if (!set.add(jsonKey)) {
-            throw InvalidGenerationSourceError(
-              'More than one field has the JSON key for name "$jsonKey".',
-              element: fe,
-            );
-          }
-          return set;
-        },
-      );
+      ..fold(<String>{}, (Set<String> set, fe) {
+        final jsonKey = nameAccess(fe);
+        if (!set.add(jsonKey)) {
+          throw InvalidGenerationSourceError(
+            'More than one field has the JSON key for name "$jsonKey".',
+            element: fe,
+          );
+        }
+        return set;
+      });
 
     if (config.createFieldMap) {
       yield createFieldMap(accessibleFieldSet);