+++ json_serializable/lib/src/json_literal_generator.dart
@@ -13,7 +13,7 @@ import 'package:source_gen/source_gen.dart';
 import 'package:source_helper/source_helper.dart';
 
 class JsonLiteralGenerator extends GeneratorForAnnotation<JsonLiteral> {
-  const JsonLiteralGenerator();
+  const JsonLiteralGenerator() : super(inPackage: 'json_annotation');
 
   @override
   Future<String> generateForAnnotatedElement(
@@ -23,12 +23,15 @@ class JsonLiteralGenerator extends GeneratorForAnnotation<JsonLiteral> {
   ) async {
     if (p.isAbsolute(annotation.read('path').stringValue)) {
       throw ArgumentError(
-          '`annotation.path` must be relative path to the source file.');
+        '`annotation.path` must be relative path to the source file.',
+      );
     }
 
     final sourcePathDir = p.dirname(buildStep.inputId.path);
-    final fileId = AssetId(buildStep.inputId.package,
-        p.join(sourcePathDir, annotation.read('path').stringValue));
+    final fileId = AssetId(
+      buildStep.inputId.package,
+      p.join(sourcePathDir, annotation.read('path').stringValue),
+    );
     final content = json.decode(await buildStep.readAsString(fileId));
 
     final asConst = annotation.read('asConst').boolValue;