+++ json_serializable/lib/src/type_helpers/config_types.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:analyzer/dart/constant/value.dart';
+import 'package:analyzer/dart/element/element.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 /// Represents values from [JsonKey] when merged with local configuration.
@@ -57,7 +58,7 @@ class ClassConfig {
   final bool genericArgumentFactories;
   final bool ignoreUnannotated;
   final bool includeIfNull;
-  final Map<String, String> ctorParamDefaults;
+  final List<FormalParameterElement> ctorParams;
   final List<DartObject> converters;
 
   const ClassConfig({
@@ -76,7 +77,7 @@ class ClassConfig {
     required this.ignoreUnannotated,
     required this.includeIfNull,
     this.converters = const [],
-    this.ctorParamDefaults = const {},
+    this.ctorParams = const [],
   });
 
   factory ClassConfig.fromJsonSerializable(JsonSerializable config) =>
@@ -89,7 +90,8 @@ class ClassConfig {
             config.createFieldMap ?? ClassConfig.defaults.createFieldMap,
         createJsonKeys:
             config.createJsonKeys ?? ClassConfig.defaults.createJsonKeys,
-        createPerFieldToJson: config.createPerFieldToJson ??
+        createPerFieldToJson:
+            config.createPerFieldToJson ??
             ClassConfig.defaults.createPerFieldToJson,
         createFactory:
             config.createFactory ?? ClassConfig.defaults.createFactory,
@@ -100,10 +102,12 @@ class ClassConfig {
             config.explicitToJson ?? ClassConfig.defaults.explicitToJson,
         includeIfNull:
             config.includeIfNull ?? ClassConfig.defaults.includeIfNull,
-        genericArgumentFactories: config.genericArgumentFactories ??
+        genericArgumentFactories:
+            config.genericArgumentFactories ??
             ClassConfig.defaults.genericArgumentFactories,
         fieldRename: config.fieldRename ?? ClassConfig.defaults.fieldRename,
-        disallowUnrecognizedKeys: config.disallowUnrecognizedKeys ??
+        disallowUnrecognizedKeys:
+            config.disallowUnrecognizedKeys ??
             ClassConfig.defaults.disallowUnrecognizedKeys,
         // TODO typeConverters = []
       );
@@ -128,20 +132,20 @@ class ClassConfig {
   );
 
   JsonSerializable toJsonSerializable() => JsonSerializable(
-        checked: checked,
-        anyMap: anyMap,
-        constructor: constructor,
-        createFactory: createFactory,
-        createToJson: createToJson,
-        createFieldMap: createFieldMap,
-        createJsonKeys: createJsonKeys,
-        createPerFieldToJson: createPerFieldToJson,
-        ignoreUnannotated: ignoreUnannotated,
-        explicitToJson: explicitToJson,
-        includeIfNull: includeIfNull,
-        genericArgumentFactories: genericArgumentFactories,
-        fieldRename: fieldRename,
-        disallowUnrecognizedKeys: disallowUnrecognizedKeys,
-        // TODO typeConverters = []
-      );
+    checked: checked,
+    anyMap: anyMap,
+    constructor: constructor,
+    createFactory: createFactory,
+    createToJson: createToJson,
+    createFieldMap: createFieldMap,
+    createJsonKeys: createJsonKeys,
+    createPerFieldToJson: createPerFieldToJson,
+    ignoreUnannotated: ignoreUnannotated,
+    explicitToJson: explicitToJson,
+    includeIfNull: includeIfNull,
+    genericArgumentFactories: genericArgumentFactories,
+    fieldRename: fieldRename,
+    disallowUnrecognizedKeys: disallowUnrecognizedKeys,
+    // TODO typeConverters = []
+  );
 }